+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.1
+- Change `deleteByShopifyId` to a hard delete. Without that, products will get trashed and conflict with new imports
+
 ## 1.2.0
 
 ### Added